requirements.txt
argcomplete==1.10.0
argon2-cffi==20.1.0
asgiref==3.3.4
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.8.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
cryptography==3.4.6
Django==3.2
django-appconf==1.0.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-crum==0.7.9
django-environ==0.4.5
django-filter==2.4.0
django-imagekit==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-storages==1.11.1
djangorestframework==3.12.2
docx==0.2.4
docx2txt==0.8
easy-thumbnails==2.7.1
EbookLib==0.17.1
extract-msg==0.23.1
idna==2.10
IMAPClient==2.1.0
importlib-metadata==3.7.3
lxml==4.6.3
Markdown==3.3.4
olefile==0.46
pdfminer.six==20181108
pilkit==2.0
Pillow==8.1.2
pinax-eventlog==5.1.0
pinax-events==3.0.0
pip==21.0.1
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodome==3.10.1
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2021.1
redis==3.5.3
requests==2.25.1
setuptools==56.0.0
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.4
wheel==0.36.2
xlrd==1.2.0
XlsxWriter==1.3.8
zipp==3.4.1
